requirements.txt
flask==3.0.3
flask-caching==1.11.1
requests==2.32.3
beautifulsoup4==4.12.3
PyPDF2==3.0.1
numpy==1.26.4
openai==1.30.1
python-dotenv==1.0.1 